requirements.txt
# huggingface-hub
llama-index==0.14.13
llama-index-embeddings-huggingface
fastmcp==2.14.5
langchain
langchain-openai
langgraph
pandas
requests
